requirements.txt
beautifulsoup4==4.12.2
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
ebaysdk==2.2.0
et_xmlfile==2.0.0
Flask==3.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
numpy==1.26.4
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
